Moved DatasetStorageLocation methods to the service
